+++ lib/rubocop/rspec/factory_bot.rb
@@ -1,0 +1,7 @@
+module RuboCop
+  module RSpec
+    # RuboCop FactoryBot project namespace
+    module FactoryBot
+    end
+  end
+end