+++ lib/rubocop/rspec/factory_girl.rb
@@ -1,7 +1,0 @@
-module RuboCop
-  module RSpec
-    # RuboCop FactoryGirl project namespace
-    module FactoryGirl
-    end
-  end
-end