+++ spec/rubocop/rspec/language/selector_set_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe RuboCop::RSpec::Language::SelectorSet do
     expect(selector_set).not_to eq(described_class.new(%i[foo bar baz]))
   end
 
-  context '#include?' do
+  describe '#include?' do
     it 'returns false for selectors not in the set' do
       expect(selector_set.include?(:baz)).to be(false)
     end
@@ -21,25 +21,25 @@ RSpec.describe RuboCop::RSpec::Language::SelectorSet do
     end
   end
 
-  context '#node_pattern' do
+  describe '#node_pattern' do
     it 'builds a node pattern' do
       expect(selector_set.node_pattern).to eql(':foo :bar')
     end
   end
 
-  context '#node_pattern_union' do
+  describe '#node_pattern_union' do
     it 'builds a node pattern union' do
       expect(selector_set.node_pattern_union).to eql('{:foo :bar}')
     end
   end
 
-  context '#send_pattern' do
+  describe '#send_pattern' do
     it 'builds a send matching pattern' do
       expect(selector_set.send_pattern).to eql('(send _ {:foo :bar} ...)')
     end
   end
 
-  context '#block_pattern' do
+  describe '#block_pattern' do
     it 'builds a block matching pattern' do
       expect(selector_set.block_pattern).to eql(
         '(block (send _ {:foo :bar} ...) ...)'