+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 * Fix exception in `RSpec/ReturnFromStub` on empty block. ([@yevhene][])
+* Fix `RSpec/NestedGroups` cop support --auto-gen-config. ([@walf443][])
 
 ## 1.19.0 (2017-10-18)
 
@@ -262,3 +263,4 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@zverok]: https:/github.com/zverok
 [@timrogers]: https://github.com/timrogers
 [@yevhene]: https://github.com/yevhene
+[@walf443]: https://github.com/walf443