+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Master (Unreleased)
 
+* Rename namespace `FactoryGirl` to `FactoryBot` following original library update. ([@walf443][])
 * Fix exception in `RSpec/ReturnFromStub` on empty block. ([@yevhene][])
 
 ## 1.19.0 (2017-10-18)
@@ -262,3 +263,4 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@zverok]: https:/github.com/zverok
 [@timrogers]: https://github.com/timrogers
 [@yevhene]: https://github.com/yevhene
+[@walf443]: https://github.com/walf443